Refactor routing utilities to avoid type issues
