+++ packages/starlight/__tests__/basics/starlight-page-route-data.test.ts
@@ -1,5 +1,5 @@
 import { expect, test, vi } from 'vitest';
-import { generateRouteData } from '../../utils/route-data';
+import { generateRouteData } from '../../utils/routing/data';
 import { routes } from '../../utils/routing';
 import {
 	generateStarlightPageRouteData,
