+++ packages/starlight/locals.ts
@@ -1,6 +1,6 @@
 import { defineMiddleware } from 'astro:middleware';
 import { useTranslations } from './utils/translations';
-import { useRouteData } from './utils/route-data';
+import { useRouteData } from './utils/routing/data';
 
 export const onRequest = defineMiddleware(async (context, next) => {
 	context.locals.t = useTranslations(context.currentLocale);
